Move the exposed Metalconfig struct to separate package
